# documentation/interface-participant-openapi.json
{
  "servers": [
    {
      "url": ""
    }
  ],
  "openapi": "3.0.3",
  "info": {
    "x-logo": {
      "url": "https://bceao.int/themes/medianet/image/png/logo_fr.png"
    },
    "title": "API exposée par le participant à l'AIP",
    "version": "2.4.0",
    "description": "Le participant qui veut utiliser l'AIP devra mettre en place une API qui lui permettra de recevoir des données provenant du système interopérable.\nVeuillez vous référer à la documentation détaillée de chaque API pour plus d'informations sur les endpoints, les paramètres et les réponses attendues.\n",
    "contact": {
      "email": "pisfn@bceao.int"
    },
    "license": {
      "name": "BCEAO",
      "url": "https://www.bceao.int"
    }
  },
...
  "paths": {
...
    "/notifications/accuse-reception": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Recevoir un accusé de réception",
        "description": "Ce point de terminaison permet de recevoir l'accusé de réception suite à un test de connectivité précédemment envoyé par le participant.\nIl nécessite un objet JSON contenant les informations de l'accusé de réception.\n\nL'API retourne une réponse HTTP 202 si l'accusé de réception a été reçu avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccuseReception"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "L'accusé de réception a été reçu avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/info-warn": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Recevoir une notification INFO ou WARN",
        "description": "Ce point de terminaison permet de recevoir des notification de type INFO ou WARN.\nIl nécessite un objet JSON contenant les informations de la notification.\n\nL'API retourne une réponse HTTP 202 si la notification a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationINFOouWARN"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La notification a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications/relation": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Recevoir une notification de relation",
        "description": "Ce point de terminaison permet de recevoir des notification de création ou de rupture de relation.\nIl nécessite un objet JSON contenant les informations de la notification.\n\nL'API retourne une réponse HTTP 202 si la notification a été reçue avec succès,\nune réponse HTTP 400 si les données en entrée sont invalides,\net une réponse HTTP 404 si l'endpoint n'a pas pu être contacté.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationRelation"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "La notification a été reçue avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Les données en entrée sont invalides.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseError"
                }
              }
            }
          },
          "404": {
            "description": "L'endpoint n'a pas pu être contacté.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReponseErrorEndpoint"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReponseSuccess": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Le message de succès",
            "example": "Demande en cours de traitement"
          }
        }
      },
      "ReponseError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Le message d'erreur",
            "example": "Les données envoyées ne sont pas conformes"
          }
        }
      },
      "ReponseErrorEndpoint": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Le message d'erreur",
            "example": "L'url n'a pas pu être contacté"
          }
        }
      },
...
      "AccuseReception": {
        "type": "object",
        "properties": {
          "msgId": {
            "pattern": "^MPIUMOA[a-zA-Z0-9]{28}$",
            "format": "M(1 position) + CodeMembre(6 positions) + UNIQUEID(28 positions)",
            "title": "MsgId",
            "type": "string",
            "description": "L'identifiant unique du message transmis depuis la plateforme interopérable.\n"
          },
          "msgIdDemande": {
            "type": "string",
            "description": "L'identifiant du message de test de connectivité auquel ce message répond."
          },
          "evenement": {
            "type": "string",
            "pattern": "[a-zA-Z0-9]{4}",
            "description": "Code proriétaire sur 4 caractères pour identifier le type d'evenement du systeme."
          },
          "evenementDescription": {
            "type": "string",
            "description": "Champ libre pour décrire l'evenement dans le systeme."
          },
          "evenementDate": {
            "type": "string",
            "description": "Date et heure de l'evenement dans le systeme"
          }
        },
        "required": [
          "msgId",
          "msgIdDemande",
          "evenement",
          "evenementDescription",
          "evenementDate"
        ],
        "example": {
          "msgId": "MPIUMOA9vydyOl7iIZSMdz3GZfWcWBa1854",
          "msgIdDemande": "MSNF0019vydyOl7iIZSMdz3GZfWcWBa1125",
          "evenement": "PING",
          "evenementDescription": "Test de connectivité",
          "evenementDate": "2024-09-21T15:30:02.050Z"
        }
      },
      "NotificationINFOouWARN": {
        "type": "object",
        "properties": {
          "msgId": {
            "pattern": "^MPIUMOA[a-zA-Z0-9]{28}$",
            "format": "M(1 position) + CodeMembre(6 positions) + UNIQUEID(28 positions)",
            "title": "MsgId",
            "type": "string",
            "description": "L'identifiant unique du message transmis depuis la plateforme interopérable.\n"
          },
          "evenement": {
            "title": "Evenement",
            "enum": [
              "INFO",
              "WARN"
            ],
            "type": "string",
            "description": "Code propriétaire sur 4 caractères pour identifier le type d'evenement du systeme.\n"
          },
          "evenementDescription": {
            "title": "EvenementDescription",
            "type": "string",
            "description": "Champ libre pour décrire l'evenement dans le systeme.\n"
          },
          "evenementDate": {
            "title": "EvenementDate",
            "format": "date-time",
            "type": "string",
            "description": "Date et heure de l'evenement dans le systeme\n"
          }
        },
        "required": [
          "msgId",
          "evenement",
          "evenementDescription",
          "evenementDate"
        ],
        "example": {
          "msgId": "MPIUMOAvydyOl7iIZSMdz3GZfWcWBa1125",
          "evenement": "INFO",
          "evenementDescription": "RECOSNB00020240801110000 DISPONIBLE",
          "evenementDate": "2024-08-29T11:01:00.000Z"
        }
      },
...
      "NotificationRelation": {
        "type": "object",
        "properties": {
          "msgId": {
            "pattern": "^MPIUMOA[a-zA-Z0-9]{28}$",
            "format": "M(1 position) + CodeMembre(6 positions) + UNIQUEID(28 positions)",
            "title": "MsgId",
            "type": "string",
            "description": "L'identifiant unique du message transmis depuis la plateforme interopérable.\n"
          },
          "codeMembreParticipant": {
            "title": "codeMembreParticipant",
            "type": "string",
            "pattern": "(BJ|BF|CI|GW|ML|NE|SN|TG)[BCDEF]{1}[0-9]{3}",
            "description": "Code membre du participant sponsor ou sponsorisé.\n"
          },
          "montantGarantiePlafond": {
            "title": "montantGarantiePlafond",
            "type": "string",
            "description": "Montant de la garantie de la sponsorisation.\n"
          },
          "dateDebut": {
            "title": "EvenementDate",
            "format": "date-time",
            "type": "string",
            "description": "Date et heure de l'evenement dans le systeme\n"
          },
          "dateFin": {
            "title": "dateFin",
            "format": "date-time",
            "type": "string",
            "description": "Date et heure de début de la relation.\n"
          }
        },
        "required": [
          "msgId",
          "codeMembreParticipant",
          "montantGarantiePlafond",
          "dateDebut"
        ],
        "example": {
          "msgId": "MPIUMOA78dyOl7iIZSMdz3GZfWcWBa4520",
          "codeMembreParticipant": "SNB000",
          "montantGarantiePlafond": "1000000",
          "dateDebut": "2024-08-29T11:01:00.000Z"
        }
      },
...
    }
  }
}